refactor: extract logic in an utils
